+++ src/main/resources/db/changelog-master.xml
@@ -16,5 +16,6 @@
     <include file="changelog-3.7.yaml" relativeToChangelogFile="true"/>
     <include file="changelog-premature.yaml" relativeToChangelogFile="true"/>
     <include file="changelog-fix-diff.yaml" relativeToChangelogFile="true"/>
+    <include file="changelog-3.8.yaml" relativeToChangelogFile="true"/>
 
 </databaseChangeLog>
