+++ src/test/java/com/gepardec/mega/service/impl/sync/SyncServiceImplTest.java
@@ -68,5 +68,4 @@ class SyncServiceImplTest {
         verify(employeeService, times(1)).getAllActiveEmployees();
         verify(userRepository, times(1)).listAll();
     }
-
 }
